requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
streamlit>=1.28.0
streamlit-cropper>=0.2.1
scipy>=1.11.0
scikit-learn>=1.3.0